+++ NotesApp/Pages/Notes/Index.cshtml
@@ -8,7 +8,8 @@
 <h1>List of Notes</h1>
 
 <p>
-    <a asp-page="Create">Create New Note</a>
+    <a asp-page="/ToDoNotes/Create">Create New ToDo List</a>
+    <a asp-page="/TextNotes/Create">Create New Text Note</a>
 </p>
 
 <form asp-page="./Index" method="get">
@@ -41,8 +42,8 @@
                             </p>
                             <div class="d-flex justify-content-between align-items-center">
                                 <div class="btn-group">                                    
-                                    <a asp-page="./Edit" asp-route-id="@note.Id" class="btn btn-sm btn-outline-secondary">Edit</a>
-                                    <a asp-page="./Delete" asp-route-id="@note.Id" class="btn btn-sm btn-outline-secondary">Delete</a>
+                                    <a asp-page="/ToDoNotes/Edit" asp-route-id="@note.Id" class="btn btn-sm btn-outline-secondary">Edit</a>
+                                    <a asp-page="/ToDoNotes/Delete" asp-route-id="@note.Id" class="btn btn-sm btn-outline-secondary">Delete</a>
                                 </div>
                                 <small class="text-muted">@note.CreationDate.ToString("dd/MM/yyyy")</small>
                             </div>
