+++ NotesApp/Pages/TextNotes/Create.cshtml
@@ -32,7 +32,7 @@
 </div>
 
 <div>
-    <a asp-page="Index">Back to List</a>
+    <a asp-page="/Notes/Index">Back to List</a>
 </div>
 
 @section Scripts {
