+++ NotesApp/Pages/TextNotes/Delete.cshtml
@@ -35,6 +35,6 @@
     <form method="post">
         <input type="hidden" asp-for="TextNote.Id" />
         <input type="submit" value="Delete" class="btn btn-danger" /> |
-        <a asp-page="./Index">Back to List</a>
+        <a asp-page="/Notes/Index">Back to List</a>
     </form>
 </div>
