+++ NotesApp/Pages/TextNotes/Edit.cshtml
@@ -36,7 +36,7 @@
 </div>
 
 <div>
-    <a asp-page="./Index">Back to List</a>
+    <a asp-page="/Notes/Index">Back to List</a>
 </div>
 
 @section Scripts {
