--- NotesApp/Pages/TextNotes/Index.cshtml.cs
@@ -1,32 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-using Microsoft.EntityFrameworkCore;
-using NotesApp.Data;
-using NotesApp.Models;
-
-namespace NotesApp.Pages.TextNotes
-{
-    public class IndexModel : PageModel
-    {
-        private readonly NotesApp.Data.NotesAppContext _context;
-
-        public IndexModel(NotesApp.Data.NotesAppContext context)
-        {
-            _context = context;
-        }
-
-        public IList<TextNote> TextNote { get;set; } = default!;
-
-        public async Task OnGetAsync()
-        {
-            if (_context.TextNotes != null)
-            {
-                TextNote = await _context.TextNotes.ToListAsync();
-            }
-        }
-    }
-}
