+++ NotesApp/Pages/ToDoItems/Create.cshtml
@@ -26,7 +26,7 @@
 </div>
 
 <div>
-    <a asp-page="/Notes/Edit" asp-route-id="@ViewData["NoteId"]">Back to Edit Page</a>
+    <a asp-page="/ToDoNotes/Edit" asp-route-id="@ViewData["NoteId"]">Back to Edit Page</a>
 </div>
 
 @section Scripts {
