+++ NotesApp/Pages/ToDoItems/Delete.cshtml
@@ -35,6 +35,6 @@
     <form method="post">
         <input type="hidden" asp-for="ToDoItem.Id" />
         <input type="submit" value="Delete" class="btn btn-danger" /> |
-        <a asp-page="/Notes/Edit" asp-route-Id="@Model.ToDoItem.NoteId">Back to Edit</a>
+        <a asp-page="/ToDoNotes/Edit" asp-route-Id="@Model.ToDoItem.NoteId">Back to Edit</a>
     </form>
 </div>
