+++ NotesApp/Pages/ToDoNotes/Create.cshtml
@@ -27,7 +27,7 @@
 </div>
 
 <div>
-    <a asp-page="Index">Back to List</a>
+    <a asp-page="/Notes/Index">Back to List</a>
 </div>
 
 @section Scripts {
