+++ NotesApp/Pages/ToDoNotes/Delete.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model NotesApp.Pages.Notes.DeleteModel
+@model NotesApp.Pages.ToDoNotes.DeleteModel
 
 @{
     ViewData["Title"] = "Delete";
@@ -29,6 +29,6 @@
     <form method="post">
         <input type="hidden" asp-for="Note.Id" />
         <input type="submit" value="Delete" class="btn btn-danger" /> |
-        <a asp-page="./Index">Back to List</a>
+        <a asp-page="/Notes/Index">Back to List</a>
     </form>
 </div>
